fix:빈 배열일 시 오류 수정, 통신 실패에도 값이 저장되던 오류 수정

+++ src/hooks/useCache.ts
@@ -20,15 +20,18 @@ const useCache = (keyword: string) => {
     }
     await fetch(`http://localhost:4000/${encode}`)
       .then((data) => {
-        const clone = data.clone();
-        cacheStorage.put(encode, clone);
-        setTimeout(() => {
-          cacheStorage.delete(encode);
-        }, 60 * 1000);
-        return data.json();
+        if (data.ok) {
+          const clone = data.clone();
+          cacheStorage.put(encode, clone);
+          setTimeout(() => {
+            cacheStorage.delete(encode);
+          }, 60 * 1000);
+          return data.json();
+        }
+        throw data.status;
       })
       .then((data) => setCacheResult(data))
-      .catch((err) => console.log(err));
+      .catch((err) => console.log(`err :${err}`));
   };
   return cacheResult;
 };
